backend/requirement.txt
cachetools==5.5.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-speech==2.30.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
rsa==4.9
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
python-dotenv==1.0.1